# contracts/openapi/identity-api-v1.yaml
openapi: 3.1.0
info:
  title: MiniPay Identity API
  version: 1.3.0
paths:
  /oauth2/token:
    post:
      operationId: exchangeAgentDelegatedToken
      summary: Exchange an Android access token for a short-lived Agent delegated token
      description: |
        RFC 8693 profile restricted to the confidential Agent client. Exactly one internal
        audience, a least-privilege tool scope, purpose and run_id are required. The subject
        token is transient and must never be logged or forwarded to the model.
      security: [{agentClientBasic: []}]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: {$ref: '#/components/schemas/AgentTokenExchangeRequest'}
      responses:
        '200':
          description: Short-lived delegated bearer token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OAuthTokenExchangeResponse'}
        '400': {description: OAuth invalid_request, invalid_grant, invalid_scope or invalid_target}
        '401': {description: OAuth invalid_client}
...
  /api/v1/auth/consumer/code/verify:
    post:
      summary: Verify a consumer login code and issue a PKCE-bound authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumerCodeVerifyRequest'
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumerAuthorizationCode'
        '400': {$ref: '#/components/responses/Problem'}
        '423': {$ref: '#/components/responses/Problem'}
        '429': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    delegatedBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentClientBasic:
      type: http
      scheme: basic
    minipayOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/oauth2/authorize
          tokenUrl: http://localhost:8081/oauth2/token
          scopes:
            identity.profile.read: Read the authenticated consumer profile
            identity.payment-authorization.write: Issue a one-time payment authorization
            identity.payment-authorization.verify: Verify a payment authorization
            identity.consumer-payment-profile.read: Read a collection recipient's safe display profile
            identity.profile.write: Initialize the authenticated consumer profile
            identity.account.manage: Resolve and manage accounts for trusted internal services
...
  schemas:
    AgentTokenExchangeRequest:
      type: object
      additionalProperties: false
      required: [grant_type, subject_token, subject_token_type, requested_token_type, audience, scope, purpose, run_id]
      properties:
        grant_type: {type: string, const: 'urn:ietf:params:oauth:grant-type:token-exchange'}
        subject_token: {type: string, minLength: 1, writeOnly: true}
        subject_token_type: {type: string, const: 'urn:ietf:params:oauth:token-type:access_token'}
        requested_token_type: {type: string, const: 'urn:ietf:params:oauth:token-type:access_token'}
        audience:
          type: string
          enum: [agent-internal, identity-internal, wallet-internal, payment-internal, commerce-internal]
        scope: {type: string, minLength: 1}
        purpose: {type: string, pattern: '^[A-Z][A-Z0-9_]{2,63}$'}
        run_id: {type: string, format: uuid}
    OAuthTokenExchangeResponse:
      type: object
      required: [access_token, issued_token_type, token_type, expires_in, scope]
      properties:
        access_token: {type: string, writeOnly: true}
        issued_token_type: {type: string, const: 'urn:ietf:params:oauth:token-type:access_token'}
        token_type: {type: string, const: Bearer}
        expires_in: {type: integer, minimum: 1, maximum: 120}
        scope: {type: string}
...
    ConsumerCodeVerifyRequest:
      type: object
      additionalProperties: false
      required: [challengeId, code, clientId, redirectUri, codeChallenge, codeChallengeMethod, deviceId]
      properties:
        challengeId: {type: string, format: uuid}
        code: {type: string, pattern: '^\d{6}$', writeOnly: true}
        clientId: {type: string, const: minipay-android}
        redirectUri: {type: string}
        codeChallenge: {type: string, minLength: 43, maxLength: 128}
        codeChallengeMethod: {type: string, const: S256}
        deviceId: {type: string, minLength: 1, maxLength: 128}
    ConsumerAuthorizationCode:
      type: object
      additionalProperties: false
      required: [authorizationCode, expiresAt, payPasswordSet, onboardingRequired, realNameStatus, realNameVerified]
      properties:
        authorizationCode: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
        payPasswordSet: {type: boolean}
        onboardingRequired: {type: boolean}
        realNameStatus: {type: string, enum: [UNVERIFIED, PROCESSING, VERIFIED, REJECTED]}
        realNameVerified: {type: boolean}
...
    Problem: {$ref: './components/problem-details-v1.yaml#/components/schemas/ProblemDetails'}
  responses:
    Problem: {description: Problem Details, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}